tests/wpt/web-platform-tests/shadow-dom/html-elements-in-shadow-trees/inert-html-elements/test-001.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_08_01_01</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#inert-html-elements">
<meta name="assert" content="HTML Elements in shadow trees: base element must behave as inert, or not part of the document tree">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
var A_08_01_01_T01 = async_test('A_08_01_01_T01', { timeout: 5000 });

A_08_01_01_T01.checkIframeContent = A_08_01_01_T01.step_func(function () {
	//remember value to check before cleaning the context (it'll destroy the iframe)
	var valueToCheck = A_08_01_01_T01.iframe.contentWindow;
	cleanContext(A_08_01_01_T01.ctx);

	assert_equals(valueToCheck, null,
		'base html element ih a shadow tree must beahve like inert one');

	A_08_01_01_T01.done();
});


A_08_01_01_T01.step(function () {

	A_08_01_01_T01.ctx = newContext();
    var d = newRenderedHTMLDocument(A_08_01_01_T01.ctx);

    //create iframe
    var iframe = document.createElement('iframe');

    iframe.src = '../../resources/blank.html';
    iframe.setAttribute('name', 'targetIframe');
    d.body.appendChild(iframe);

    A_08_01_01_T01.iframe = iframe;

    // create a link
    var link = d.createElement('a');
    link.setAttribute('href', '../../resources/bobs_page.html');
    link.innerHTML = 'the link';
    d.body.appendChild(link);

    //create Shadow root
    var root = d.createElement('div');
    d.body.appendChild(root);
    var s = root.createShadowRoot();

    // create base element, set iframe as a target
    var base = d.createElement('base');
    base.setAttribute('target', 'targetIframe');
    s.appendChild(base);

    //click the link
    link.click();

    //Expected: base should be inert therefore document d
    // should be reloaded, so iframe context shouldn't be affected

    // set timeout to give the iframe time to load content
    setTimeout('A_08_01_01_T01.checkIframeContent()', 2000);
});
</script>
</body>
</html>
